refactor(GuidePage): 안내 로직 분리 - useGuidanceLogic 훅으로 이동
- getReachableGuidancePoint, checkArrival 함수를 독립적인 훅으로 분리
- reachedPointIndexRef 상태를 훅 내부로 캡슐화
- 코드 가독성 및 재사용성 향상
- index.tsx 13줄 감소 (391줄 → 378줄)
